sort bower

+++ app/scripts/controllers/make.js
@@ -76,7 +76,9 @@ angular.module('bitbloqApp')
 
                 if ($scope.project._id) {
                     if (!$scope.project._acl || ($scope.project._acl['user:' + $scope.common.user._id] && $scope.project._acl['user:' + $scope.common.user._id].permission === 'ADMIN')) {
-                        if ($scope.tempImage.blob) {
+                        if ($scope.tempImage.blob && $scope.tempImage.blob.size) {
+
+                            console.log('tengo imagen :) *************');
                             if (currentProject.imageType !== $scope.tempImage.blob.type) {
                                 imageApi.delete($scope.project._id, currentProject.imageType);
                             }
@@ -116,7 +118,9 @@ angular.module('bitbloqApp')
                 } else {
                     if ($scope.common.user) {
                         currentProject.creatorId = $scope.project.creatorId = $scope.common.user._id;
-                        if ($scope.tempImage.blob) {
+                        if ($scope.tempImage.blob && $scope.tempImage.blob.size) {
+
+                            console.log('tengo imagen :) *************');
                             currentProject.imageType = $scope.tempImage.blob.type;
                         }
                         projectApi.save(currentProject).then(function(response) {
@@ -564,7 +568,7 @@ angular.module('bitbloqApp')
 
         $scope.projectHasChanged = function(currentProject, oldProject) {
             var identicalProjectObject = _.isEqual(currentProject, oldProject);
-            return !identicalProjectObject || $scope.tempImage.blob;
+            return !identicalProjectObject ||  ($scope.tempImage.blob && $scope.tempImage.blob.size);
         };
 
         $scope.updateBloqs = function() {
